Patch notes - Renaming some files to a properly name. - Refactor of the custom hook that was made to connect the components to the store. - Refactor of some styled components. - Refactor of the Loader logic. - Refactor of the redux store and sagas.
