+++ src/App.style.js
@@ -1,4 +1,4 @@
-import styled, { createGlobalStyle, css } from "styled-components/macro";
+import { createGlobalStyle, css } from "styled-components/macro";
 
 export const GlobalStyles = createGlobalStyle`
   * {
@@ -14,32 +14,6 @@ export const GlobalStyles = createGlobalStyle`
 body {
   overflow: hidden;
 }
-`;
-
-export const Button = styled.button`
-  background: transparent;
-  border-radius: 3px;
-  border: 2px solid orange;
-  color: orange;
-  padding: 0.25em 1em;
-  cursor: pointer;
-
-  ${({ width }) =>
-    width
-      ? `
-			width: ${width}`
-      : `min-width: 65px;`};
-
-  ${({ fontSize }) => fontSize && `font-size: ${fontSize}`};
-
-  ${({ stock }) => {
-    if (stock === 0) {
-      return `
-      color: #e8e8e8;
-      border: 2px solid #e8e8e8;
-      `;
-    }
-  }}
 `;
 
 export const VerticalCenter = css`
