+++ src/components/Header/Header.style.js
@@ -1,0 +1,7 @@
+import styled from "styled-components";
+
+export const Header = styled.div`
+  display: flex;
+  flex-flow: column;
+  height: 100vh;
+`;
