+++ src/hocs/withHeader.js
@@ -1,17 +1,15 @@
 import React from "react";
+import { Header } from "../components/Header/Header.style";
 import NavBar from "../components/NavBar/NavBar";
 
 function withHeader(Component, options) {
   const { pageName = "" } = options;
   function withHeaderWrapper({ ...props }) {
     return (
-      <div
-        className={"headerContainer"}
-        style={{ display: "flex", flexFlow: "column", height: "100vh" }}
-      >
+      <Header>
         <NavBar pageName={pageName} />
         <Component {...props} />
-      </div>
+      </Header>
     );
   }
