+++ src/hooks/useProduct.js
@@ -2,26 +2,20 @@ import { useCallback } from "react";
 import { useDispatch, useSelector } from "react-redux";
 
 import {
-  addProductToCartAction,
   loadFavoriteProductsAction,
   loadMoreProductsAction,
   loadProductsAction,
-  removeProductFromCartAction,
-  updateCartProductAction,
   updateProductAction,
 } from "../store";
 
-/** Custom hook for accessing redux store */
-function useProducts() {
+/** Custom hook for accessing "Product" state from redux store */
+function useProduct() {
   const dispatch = useDispatch();
 
   return {
     // Selectors
-    loading: useSelector(
-      (state) => state.checkout.loading || state.cart.loading
-    ),
-    products: useSelector((state) => state.checkout.data),
-    cart: useSelector((state) => state.cart.data),
+    loading: useSelector((state) => state.products.loading),
+    products: useSelector((state) => state.products.data),
     // Dispatchers
     getProducts: useCallback(() => dispatch(loadProductsAction()), [dispatch]),
     getMoreProducts: (page) => dispatch(loadMoreProductsAction(page)),
@@ -29,11 +23,8 @@ function useProducts() {
       () => dispatch(loadFavoriteProductsAction()),
       [dispatch]
     ),
-    addProduct: (product) => dispatch(addProductToCartAction(product)),
     updateProduct: (product) => dispatch(updateProductAction(product)),
-    updateCartProduct: (product) => dispatch(updateCartProductAction(product)),
-    removeProduct: (id) => dispatch(removeProductFromCartAction(id)),
   };
 }
 
-export default useProducts;
+export default useProduct;
