+++ src/store/index.js
@@ -1,15 +1,16 @@
 import { combineReducers } from 'redux';
-import { cartReducer } from './cart.reducer';
-import { checkoutReducer } from './checkout.reducers';
+import { cartReducer } from './reducers/cart.reducer';
+import { productsReducer } from './reducers/products.reducers';
 
-export * from './checkout.actions';
-export * from './checkout.reducers';
-export * from './cart.actions';
-export * from './cart.reducer';
-export * from './store.saga';
+export * from './actions/products.actions';
+export * from './reducers/products.reducers';
+export * from './actions/cart.actions';
+export * from './reducers/cart.reducer';
+export * from './sagas/cart.saga';
+export * from './sagas/products.saga'
 
 const store = combineReducers({
-  checkout: checkoutReducer,
+  products: productsReducer,
   cart: cartReducer
 });
