+++ src/store/sagas/products.saga.js
@@ -4,18 +4,7 @@ import {
   loadMoreProductsAPI,
   loadProductsAPI,
   updateProductAPI,
-} from "./api/checkout.api";
-import {
-  ADD_PRODUCT_TO_CART,
-  ADD_PRODUCT_TO_CART_ERROR,
-  ADD_PRODUCT_TO_CART_SUCCESS,
-  REMOVE_PRODUCT_FROM_CART,
-  REMOVE_PRODUCT_FROM_CART_ERROR,
-  REMOVE_PRODUCT_FROM_CART_SUCCESS,
-  UPDATE_CART_PRODUCT,
-  UPDATE_CART_PRODUCT_ERROR,
-  UPDATE_CART_PRODUCT_SUCCESS,
-} from "./cart.actions";
+} from "../api/products.api";
 import {
   LOAD_FAVORITE_PRODUCTS,
   LOAD_MORE_PRODUCTS,
@@ -27,7 +16,7 @@ import {
   UPDATE_PRODUCT,
   UPDATE_PRODUCT_ERROR,
   UPDATE_PRODUCT_SUCCESS,
-} from "./checkout.actions";
+} from "../actions/products.actions";
 
 export function* loadingProductAsync() {
   try {
@@ -88,60 +77,11 @@ export function* watchUpdatingProductAsync() {
   yield takeEvery(UPDATE_PRODUCT, updatingProductAsync);
 }
 
-export function* addingProductToCartAsync({ payload }) {
-  try {
-    const data = yield call(updateProductAPI, payload);
-    const updatedProduct = data;
-    yield put({ type: UPDATE_PRODUCT_SUCCESS, payload: updatedProduct });
-    yield put({ type: ADD_PRODUCT_TO_CART_SUCCESS, payload: updatedProduct });
-  } catch (err) {
-    yield put({ type: ADD_PRODUCT_TO_CART_ERROR, payload: err });
-  }
-}
-
-export function* watchAddingProductToCartAsync() {
-  yield takeEvery(ADD_PRODUCT_TO_CART, addingProductToCartAsync);
-}
-
-export function* removeProductFromCartAsync({ payload }) {
-  try {
-    yield put({ type: REMOVE_PRODUCT_FROM_CART_SUCCESS, payload });
-  } catch (err) {
-    yield put({ type: REMOVE_PRODUCT_FROM_CART_ERROR, payload: err });
-  }
-}
-
-export function* watchRemovingProductFromCartAsync() {
-  yield takeEvery(REMOVE_PRODUCT_FROM_CART, removeProductFromCartAsync);
-}
-
-export function* updatingCartProductAsync({ payload }) {
-  try {
-    const cartProduct = { ...payload };
-
-    delete payload.quantity;
-
-    const data = yield call(updateProductAPI, payload);
-    const updatedProduct = data;
-    yield put({ type: UPDATE_PRODUCT_SUCCESS, payload: updatedProduct });
-    yield put({ type: UPDATE_CART_PRODUCT_SUCCESS, payload: cartProduct });
-  } catch (err) {
-    yield put({ type: UPDATE_CART_PRODUCT_ERROR, payload: err });
-  }
-}
-
-export function* watchUpdatingCartProductAsync() {
-  yield takeEvery(UPDATE_CART_PRODUCT, updatingCartProductAsync);
-}
-
-export function* storeSaga() {
+export function* productsSaga() {
   yield all([
     watchLoadingProductsAsync(),
     watchUpdatingProductAsync(),
-    watchAddingProductToCartAsync(),
     watchLoadingMoreProductsAsync(),
     watchLoadingFavoriteProductsAsync(),
-    watchRemovingProductFromCartAsync(),
-    watchUpdatingCartProductAsync(),
   ]);
 }
